Create Data scraping IMDB
This project is about data scraping the IMDB website and writing the different criterias in different files. The specific section the data is being written about is the IMDB top 200 shows by rating. The data is written in four different criterias. One file is all the top 200 shows, one is shows before the year 2000, one after the year 2000, and one is for shows with a rating of 8.7 or higher